Replaced Ineni by the version edited by M. Bilgin.
The previous version of Ineni was a copy of Urk. IV 53-62,
including the spacing and line skips, which correspond to Sethe's
reading of the text, and not to the original layout.

M. Bilgin has produced a version which is closer to the original, which I chose as the default version now.